+++ llvm/include/llvm/CodeGen/StackProtector.h
@@ -26,7 +26,6 @@
 namespace llvm {
 
 class BasicBlock;
-class DominatorTree;
 class Function;
 class Instruction;
 class Module;