+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -165,6 +165,11 @@ New checks
   Checks that all implicit and explicit inline functions in header files are
   tagged with the ``LIBC_INLINE`` macro.
 
+- New :doc:`misc-header-include-cycle
+  <clang-tidy/checks/misc/header-include-cycle>` check.
+
+  Check detects cyclic ``#include`` dependencies between user-defined headers.
+
 - New :doc:`misc-include-cleaner
   <clang-tidy/checks/misc/include-cleaner>` check.